requirements.txt
Django==5.0.9
psycopg[binary]==3.2.3
djangorestframework==3.15.2
djangorestframework-gis==1.1
django-environ==0.11.2
django-htmx==1.19.0
gunicorn==23.0.0
PyYAML>=6.0
requests==2.32.3
pyproj==3.7.0
python-dateutil==2.9.0.post0
whitenoise==6.7.0
